backend/requirements.txt
# Nusantara Ekspor - Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.22.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic[email-validator]==2.10.4
httpx==0.28.1
google-generativeai==0.8.4
google-genai>=1.0.0
websockets==14.1
